Add periscope error test cases
